# refactor(testing): Re-enable support shim, and extend to test options object not being modified v2