--- a/lib/model.js
+++ b/lib/model.js
@@ -571,7 +571,7 @@ class Model {
     });
   }
 
-  static init(attributes, options, modelManager) {
+  static init(attributes, options, modelManager) { // testhint options:none
 
     this.options = Utils._.extend({
       timestamps: true,