--- a/test/integration/include/schema.test.js
+++ b/test/integration/include/schema.test.js
@@ -15,7 +15,6 @@ var sortById = function(a, b) {
 
 describe(Support.getTestDialectTeaser('Includes with schemas'), function() {
   describe('findAll', function() {
-    this.timeout(30000);
     beforeEach(function() {
       var self = this;
       this.fixtureA = function() {