--- a/test/supportShim.js
+++ b/test/supportShim.js
@@ -2,7 +2,8 @@
 
 var QueryInterface = require(__dirname + '/../lib/query-interface')
   , hintsModule = require('hints')
-  , _ = require('lodash');
+  , _ = require('lodash')
+  , util = require('util');
 
 /*
  * Shims all Sequelize methods to test for logging passing.
@@ -10,20 +11,24 @@ var QueryInterface = require(__dirname + '/../lib/query-interface')
  */
 module.exports = function(Sequelize) {
   // Shim all Sequelize methods
-  shimAll(Sequelize.prototype);
-  shimAll(Sequelize.Model);
-  shimAll(Sequelize.Model.prototype);
-  shimAll(QueryInterface.prototype);
+  shimAll(Sequelize.prototype, 'Sequelize#');
+  shimAll(Sequelize.Model, 'Model.');
+  shimAll(Sequelize.Model.prototype, 'Model#');
+  shimAll(QueryInterface.prototype, 'QueryInterface#');
+  shimAll(Sequelize.Association.prototype, 'Association#');
+  _.forIn(Sequelize.Association, function(Association, name) {
+    shimAll(Association.prototype, 'Association.' + name + '#');
+  });
 
-  // Shim Model.prototype to then shim getter/setter methods
+  // Shim Model static methods to then shim getter/setter methods
   ['hasOne', 'belongsTo', 'hasMany', 'belongsToMany'].forEach(function(type) {
-    shimMethod(Sequelize.Model.prototype, type, function(original) {
+    shimMethod(Sequelize.Model, type, function(original) {
       return function(targetModel, options) {
         var model = this,
           association = original.apply(this, arguments);
 
-        _.forIn(association.accessors, function(accessor) {
-          shim(model.prototype, accessor, model.prototype[accessor].length);
+        _.forIn(association.accessors, function(accessor, accessorName) {
+          shim(model.prototype, accessor, model.prototype[accessor].length, null, 'Model#' + accessorName);
         });
 
         return association;
@@ -36,11 +41,12 @@ module.exports = function(Sequelize) {
   /*
    * Shims all shimmable methods on obj.
    * @param {Object} obj
+   * @param {String} objName Name of object for error reporting
    */
-  function shimAll(obj) {
-    _.forIn(obj, function(method, name) {
+  function shimAll(obj, objName) {
+    forOwn(obj, function(method, name) {
       var result = examine(method, name);
-      if (result) shim(obj, name, result.index, result.conform);
+      if (result) shim(obj, name, result.index, result.conform, objName + name);
     });
   }
 
@@ -72,10 +78,15 @@ module.exports = function(Sequelize) {
    * @returns {Object}
    */
   function examine(method, name) {
+    // skip if not a function
     if (typeof method !== 'function') return;
 
+    // skip classes, constructors and private methods
+    if (name === 'constructor' || !name.match(/^[a-z]/)) return;
+
     // find test hints if provided
-    var obj = hintsModule.full(method.toString(), 'testhint', {function: true}),
+    var fnStr = getFunctionCode(method),
+      obj = hintsModule.full(fnStr, 'testhint'),
       hints = obj.hints,
       tree = obj.tree;
 
@@ -95,9 +106,6 @@ module.exports = function(Sequelize) {
       return result;
     }
 
-    // skip if function name does not start with lower case letter
-    if (!name.match(/^[a-z]/)) return;
-
     // find 'options' argument - if none, then skip
     var index = args.indexOf('options');
     if (index === -1) return;
@@ -116,33 +124,60 @@ module.exports = function(Sequelize) {
    * @param {String} name Name of method on object to shim
    * @param {Integer} index Index of argument which is `options` (1-based)
    * @param {Function} conform Function to conform function arguments
+   * @param {String} debugName Full name of method for error reporting
    */
-  function shim(obj, name, index, conform) {
+  function shim(obj, name, index, conform, debugName) {
     index--;
 
     shimMethod(obj, name, function(original) {
+      var sequelizeProto = (obj === Sequelize.prototype);
+
       return function() {
+        var sequelize = (sequelizeProto ? this : this.sequelize);
+        if (this instanceof Sequelize.Association) sequelize = this.target.sequelize;
+        if (!sequelize) throw new Error('Object does not have a `sequelize` attribute');
+
         var args = Sequelize.Utils.sliceArgs(arguments),
           fromTests = calledFromTests();
 
         if (conform) args = conform.apply(this, arguments);
 
+        var options = args[index];
+
         if (fromTests) {
-          args[index] = addLogger(args[index]);
+          args[index] = options = addLogger(options, sequelize);
         } else {
-          testLogger(args[index]);
+          testLogger(options, debugName);
         }
 
-        var result;
-        // NB next line written as a single statement to avoid bug with uncaught rejection
-        return (result = original.apply(this, args)) instanceof Sequelize.Promise ?
-          result.finally(finish) :
-          finish();
+        var originalOptions = cloneOptions(options);
 
-        function finish() {
-          if (fromTests) removeLogger(args[index]);
-          return result;
+        var result = original.apply(this, args);
+
+        if (result && typeof result.then === 'function') {
+          var err;
+          try {
+            checkOptions(options, originalOptions, debugName);
+          } catch (e) {
+            err = e;
+          }
+
+          if (!(result instanceof Sequelize.Promise)) {
+            result = Sequelize.Promise.resolve(result);
+            err = new Error('Promise returned by ' + debugName + ' is not instance of Sequelize.Promise');
+          }
+
+          result = result.finally(function() {
+            if (err) throw err;
+            checkOptions(options, originalOptions, debugName);
+            if (fromTests) removeLogger(options);
+          });
+        } else {
+          checkOptions(options, originalOptions, debugName);
+          if (fromTests) removeLogger(options);
         }
+
+        return result;
       };
     });
   }
@@ -174,17 +209,20 @@ module.exports = function(Sequelize) {
    * @param {Object} options
    * @returns {Object} Options with `logging` attribute added
    */
-  function addLogger(options) {
+  function addLogger(options, sequelize) {
     if (!options) options = {};
 
     var hadLogging = options.hasOwnProperty('logging'),
       originalLogging = options.logging;
 
-    options.logging = function(msg) {
-      if (originalLogging) {
-        return originalLogging.apply(this, arguments);
-      } else {
-        logger(msg);
+    options.logging = function() {
+      var logger = originalLogging !== undefined ? originalLogging : sequelize.options.logging;
+      if (logger) {
+        if ((sequelize.options.benchmark || options.benchmark) && logger === console.log) {
+          return logger.call(this, arguments[0] + ' Elapsed time: ' + arguments[1] + 'ms');
+        } else {
+          return logger.apply(this, arguments);
+        }
       }
     };
 
@@ -201,12 +239,10 @@ module.exports = function(Sequelize) {
    * @returns {Object} Options with `logging` attribute reverted to original value
    */
   function removeLogger(options) {
-    if (options.logging && options.logging.__testLoggingFn) {
-      if (options.logging.hasOwnProperty('__originalLogging')) {
-        options.logging = options.logging.__originalLogging;
-      } else {
-        delete options.logging;
-      }
+    if (options.logging.hasOwnProperty('__originalLogging')) {
+      options.logging = options.logging.__originalLogging;
+    } else {
+      delete options.logging;
     }
   }
 
@@ -216,8 +252,8 @@ module.exports = function(Sequelize) {
    * @param {Object} options
    * @throws {Error} Throws if `options.logging` is not a shimmed logging function
    */
-  function testLogger(options) {
-    if (!options || !options.logging || !options.logging.__testLoggingFn) throw new Error('options.logging has been lost');
+  function testLogger(options, name) {
+    if (!options || !options.logging || !options.logging.__testLoggingFn) throw new Error('options.logging has been lost in method ' + name);
   }
 
   /*
@@ -232,18 +268,41 @@ module.exports = function(Sequelize) {
   function calledFromTests() {
     return !!((new Error()).stack.split(/[\r\n]+/)[3].match(regExp));
   }
-
-  /*
-   * Logging function
-   *
-   * @param {String} msg Logging message
-   */
-  function logger(msg) {
-    if (process.env.SEQ_LOG) console.log(msg);
-  }
 };
 
 // Helper functions for examining code for hints
+
+/*
+ * Loop through own properties of object (including non-enumerable properties)
+ * and call `fn` for each property with argments `(value, key, object)`.
+ * Getters are skipped.
+ * Like `_.forIn()` except also includes non-enumarable properties, and skips getters.
+ *
+ * @param {Object} obj - Object to iterate over
+ * @param {Function} fn - Function to call for each property
+ * @returns {Object} - `obj` input
+ */
+function forOwn(obj, fn) {
+  Object.getOwnPropertyNames(obj).forEach(function(key) {
+    if (Object.getOwnPropertyDescriptor(obj, key).hasOwnProperty('value')) fn(obj[key], key, obj);
+  });
+  return obj;
+}
+
+/*
+ * Get code of function
+ * Adds 'function ' to start of code where fn has been defined with object method shortcut,
+ * and alters illegal function names ('import', 'delete'), so code can be parsed by `acorn`.
+ *
+ * @param {Function} fn - Function
+ * @returns {string} - Code of function
+ */
+function getFunctionCode(fn) {
+  var code = fn.toString();
+  if (code.match(/^function[\s\*\(]/) || code.match(/^class[\s\{]/)) return code;
+  if (code.match(/^(import|delete)[\s\*\(]/)) code = '_' + code.substr(1);
+  return 'function ' + code;
+}
 
 /*
  * Returns arguments of a function as an array, from it's AST
@@ -277,8 +336,45 @@ function getArgumentsConformFn(method, args, hints, tree) {
 
   // extract
   var start = hints.start ? hints.start.end : tree.body[0].body.start + 1,
-    body = method.toString().slice(start, hints.end.start);
+    body = getFunctionCode(method).slice(start, hints.end.start);
 
   // create function that conforms arguments
   return new Function(args, body + ';return [' + args + '];'); // jshint ignore:line
+}
+
+/*
+ * Clone options object
+ * @params {Object} options - Options object
+ * @returns {Object} Clone of options
+ */
+function cloneOptions(options) {
+  return _.cloneDeepWith(options, function(value) {
+    if (typeof value === 'object' && !_.isPlainObject(value)) return value;
+  });
+}
+
+/*
+ * Checks options object has not been altered and throw if altered
+ *
+ * @params {Object} options - Options object
+ * @params {Object} original - Original options object
+ * @throws {Error} Throws if options and original are not identical
+ */
+function checkOptions(options, original, name) {
+  if (!optionsEqual(options, original)) throw new Error('options modified in ' + name + ', input: ' + util.inspect(original) + ' output: ' + util.inspect(options));
+}
+
+/*
+ * Compares two options objects and returns if they are deep equal to each other.
+ * Objects which are not plain objects (e.g. Models) are compared by reference.
+ * Everything else deep-compared by value.
+ *
+ * @params {Object} options - Options object
+ * @params {Object} original - Original options object
+ * @returns {Boolean} true if options and original are same, false if not
+ */
+function optionsEqual(options, original) {
+  return _.isEqualWith(options, original, function(value1, value2) {
+    if ((typeof value1 === 'object' && !_.isPlainObject(value1)) || (typeof value2 === 'object' && !_.isPlainObject(value2))) return value1 === value2;
+  });
 }