@janmeier 2016-05-03
@overlookmotel Travis is done now - seems to be a single place left where options are modified in postgres