@overlookmotel 2016-05-03
Travis keeps failing randomly - in different places and different dialects each time, but always due to a test timing out. Is this normal? Or is the cause specific to this PR - maybe if it's slowed things down,