@janmeier 2016-05-03
We do have periodic failures from time to time - as long as its just a single build were usually fine. But these changes have made builds take significantly longer .. :( 4-7 minutes, as opposed to 3-5 minutes per job now. I wouldn't have thought that cloning adds that amount of time, but apparently it does.

Right now all jobs are failing though - I merged a PR earlier today which used `$optClone` in a new place, so that needs to be changed to `Utils.cloneDeep`