@overlookmotel 2016-05-03
OK, have rebased on master and replaced `$optClone()` with `Utils.cloneDeep()`. Let's see how it fares this time...