@overlookmotel 2016-05-04
I increased the tests timeout to 30secs and hey presto Travis passes!

It does seem that the options cloning (or the re-introduction of the shim in general) is slowing things down a fair bit.

I think this needs further investigation before merging, but can we possibly conclude first whether the code I've built on top of @janmeier's PR is good in principle, aside from the concern about slowdown?