@janmeier 2016-05-06
The code LGTM! :)