@overlookmotel 2016-05-06
@janmeier Thanks!

But let me run a few benchmarks before merging. I want to check whether the slowdown is caused by all the options cloning in sequelize methods, or by the test shim itself. i.e. whether it's going to be slow in production, or just slow in the tests. I'd say don't merge for now until it's clear that these changes aren't massively impacting performance in production.