@overlookmotel 2016-05-11
I've split this PR into 3 parts:

https://github.com/sequelize/sequelize/pull/5874 Clones options
https://github.com/sequelize/sequelize/pull/5836 (this PR) Tests for options cloning and re-enables the support shim
https://github.com/sequelize/sequelize/pull/5875 Removes tests which are now redundant

Each PR depends on the one before so would need to be merged in the above order.

I've split it for 2 reasons:
1. To make a clearer commit history
2. To see how long tests take on Travis with options cloning but without the test shims vs with the test shims