@janmeier 2016-05-13
Good idea splitting - 2 needs a rebase now