@overlookmotel 2016-05-16
Have rebased this onto master after merge of PR https://github.com/sequelize/sequelize/pull/5874.

Enabling the tests shim approximately doubles the run time of the tests.

On my development machine (MacBook Pro Retina), the times to run MySQL integration tests:
Without options cloning: 99 secs
With options cloning: 101 secs
With test shim: 198 secs
(the above are averages over 5 runs of each)

The above is also roughly borne out by Travis's timings.
Without options cloning: 24m 4s
With options cloning: 17m 41s
With test shim: 34m 35s
(Travis's timings seem to diverge between runs quite widely, so the above is only a rough indicator).

So... question is: is it worth it?

In my opinion, having universal tests which check for passing of `options.logging` and that `options` object are not modified is a pretty good thing. Also, I'd now like to add to the shim to test for a few other cases:

1. `options.transaction` passed correctly
2. CLS context maintained correctly
3. All Promises returned from Sequelize methods are instances of `Sequelize.Promise` (this is important for CLS)

`options.transaction` and CLS are both in my opinion critical, so I think this would increase the utility of the test shim a lot.

Also in my personal opinion, the long-term solution to decreasing build times is to move to running tests in parallel (Mocha 3 will apparently support this https://github.com/mochajs/mocha/issues/1969 or could try [mocha.parallel](https://github.com/danielstjules/mocha.parallel)). Reducing number of supported Node versions would also help. But maybe all of these solutions are a long way off.

@mickhansen I guess you'd be the one most affected by the increased test run times. What's your opinion? Is it worth it?