@sushantdhiman 2016-05-16
> With test shim: 34m 35s

Test shims are cool but thats too much time. Ideally we should reduce (or remove) IO js builds if we want use these updates.