@janmeier 2016-05-28
Since we have cut down the test times considerably, I think we should add this - Its important to make sure options always work. Perhaps we could enable it for one