@overlookmotel 2016-06-11
I've rebased this on master and adapted it for the change to ES6 classes.

However, whether we can deal with the longer build times remains an open question.

So I've disabled the shims for now by default. To run them you have to use `SHIM=true npm test`. So this  PR won't affect the build times on Travis.

Could we possibly get this merged at this point just so the shims code doesn't get behind the main codebase again? We can discuss later whether it's worth enabling the shims or not. And in the meantime anyone who wants to run them can do so locally.

How's that sound for a compromise?