@janmeier 2016-06-12
Sounds like a good compromise to me. Perhaps we can run the shims on one node version only, like with coverage.

We could also set it to run pre-release, but that would hinder us from doing quick fix-releases