@overlookmotel 2016-06-12
@janmeier Thanks for merging. I guess we can come back to the question of whether to enable the shims later. Obviously I'd like to see them enabled! But I can totally see that long build times hinder productivity.